pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-checkpoint-delta"
version = "0.1.0"
description = "A LangGraph checkpointer that persists agent state to Delta Lake tables via Unity Catalog"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.9"
keywords = [
    "langgraph",
    "delta-lake",
    "checkpointer",
    "databricks",
    "unity-catalog",
    "agents",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core deps — only what's needed regardless of backend
dependencies = [
    "langgraph-checkpoint>=2.0.0",
]

[project.optional-dependencies]
# Rust-backed Delta writer — no Spark required, works anywhere
deltalake = [
    "deltalake>=0.19.0",
]
# Spark-based Delta writer — required for Databricks Unity Catalog writes via Spark
spark = [
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
]
# Install both backends
all = [
    "deltalake>=0.19.0",
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
]
# Dev/test dependencies
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "mypy>=1.8",
    "ruff>=0.3",
    "langgraph>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/langgraph-checkpoint-delta"
Repository = "https://github.com/yourusername/langgraph-checkpoint-delta"
Issues = "https://github.com/yourusername/langgraph-checkpoint-delta/issues"

[tool.setuptools.packages.find]
include = ["langgraph_checkpoint_delta*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true